+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changelog
  * renamed `LocalResource` to `AbstractLocalResource`
  * added `AbstractResource`
  * `CompositeRepository` now sets the correct path for returned resources
+ * moved `Selector` to `Puli\Repository\Selector` namespace
 
 * 1.0.0-alpha4 (2014-12-03)